Refactor getUrl into separate file. Ensure that only .amp is stripped from URL (regex was too greedy)
