+++ src/app/routes/utils/fetchPageData/utils/getUrl/index.js
@@ -1,0 +1,17 @@
+import { getQueryString, getUrlPath } from '#lib/utilities/urlParser';
+import getBaseUrl from '../getBaseUrl';
+import onClient from '#lib/utilities/onClient';
+import isLive from '#lib/utilities/isLive';
+
+const baseUrl = onClient()
+  ? getBaseUrl(window.location.origin)
+  : process.env.SIMORGH_BASE_URL;
+
+export default pathname => {
+  if (!pathname) return '';
+
+  const ampRegex = /\.amp$/;
+  const params = isLive() ? '' : getQueryString(pathname);
+  const basePath = getUrlPath(pathname);
+  return `${baseUrl}${basePath.replace(ampRegex, '')}.json${params}`; // Remove .amp at the end of pathnames for AMP pages.
+};
